+++ src/utils/constants.ts
@@ -1,1 +1,2 @@
-export const stripeApiUrl = "https://api.stripe.com";
+export const stripeApiUrl = "https://api.stripe.com/v1";
+export const stripeApiVersion = "2020-08-27";